Hotfix 2

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -10,7 +10,7 @@
 
 size_t print_dlistint(const dlistint_t *h)
 {
-	unsigned int n;
+	size_t n;
 
 	for (n = 0; h; n++)
 	{
